add useTokenBalance
