+++ src/pages/Root/Administrator.tsx
@@ -12,15 +12,16 @@ import {useRecoilValue} from "recoil";
 import {activeChannelIdAtom} from "../../state/Root";
 import {useActiveChannelInfo} from "../../hooks/useActiveChannelInfo";
 import {useActiveWeb3React} from "../../hooks/web3";
-import {FC, useCallback, useState} from "react";
-import {ERROR, IDLE, IDLE_DELAY, PROCESSING, SUCCESS, ZERO_ADDRESS} from "../../constants/misc";
+import {FC, useState} from "react";
+import {ERROR, IDLE, IDLE_DELAY, PROCESSING, SUCCESS} from "../../constants/misc";
 import {useNestOpenPlatformContract, useTokenContract} from "../../hooks/useContract";
 import {NEST_OPEN_PLATFORM_ADDRESS} from "../../constants/addresses";
-import {formatNumber, parseToBigNumber} from "../../utils/bignumberUtil";
+import {parseToBigNumber} from "../../utils/bignumberUtil";
 import {useTokenSymbol} from "../../hooks/Tokens";
 import {formatWithUnit, parseToNumber} from "../../utils/unit";
 import {useBalance} from "../../hooks/useBalance";
 import {CHAIN_INFO} from "../../constants/chains";
+import {useTokenBalance} from "../../hooks/useTokenBalance";
 
 const Administrator = () => {
   const {info, status} = useActiveChannelInfo()
@@ -47,34 +48,22 @@ const Administrator = () => {
   )
 }
 
-type PopverProps = {
+type PopoverProps = {
   isLoading: boolean
   tokenAddress?: string
 }
 
-const DepositPopover: FC<PopverProps> = ({...props}) => {
-  const {chainId} = useActiveWeb3React()
+const DepositPopover: FC<PopoverProps> = ({...props}) => {
+  const {chainId, account} = useActiveWeb3React()
   const activeChannelId = useRecoilValue(activeChannelIdAtom)
   const nestOpenPlatform = useNestOpenPlatformContract(NEST_OPEN_PLATFORM_ADDRESS[chainId ?? 1], true)
   const token = useTokenContract(props.tokenAddress)
-  const {account} = useActiveWeb3React()
-  const [balance, setBalance] = useState('0')
+  const balance = useTokenBalance(props.tokenAddress, account)
   const [amount, setAmount] = useState('0')
   const tokenSymbol = useTokenSymbol(props.tokenAddress ?? "")
   const [depositStatus, setDepositStatus] = useState(IDLE)
   const [approveStatus, setApproveStatus] = useState(IDLE)
   const {refresh: refreshChannelInfo} = useActiveChannelInfo()
-
-  const refresh = useCallback(async () => {
-    if (!token) return
-    try {
-      const res = await token.balanceOf(account ?? ZERO_ADDRESS)
-      setBalance(formatNumber(parseToBigNumber(res).shiftedBy(-18)))
-    }catch (e){
-      setBalance('NaN')
-    }
-  }, [account, token])
-  setImmediate(refresh, 3000)
 
   const handleDeposit = async () => {
     if (!nestOpenPlatform) return
@@ -183,28 +172,15 @@ const DepositPopover: FC<PopverProps> = ({...props}) => {
   )
 }
 
-const WithdrawPopover: FC<PopverProps> = ({...props}) => {
+const WithdrawPopover: FC<PopoverProps> = ({...props}) => {
   const {chainId, account} = useActiveWeb3React()
   const activeChannelId = useRecoilValue(activeChannelIdAtom)
   const nestOpenPlatform = useNestOpenPlatformContract(NEST_OPEN_PLATFORM_ADDRESS[chainId ?? 1], true)
   const [amount, setAmount] = useState('0')
-  const token = useTokenContract(props.tokenAddress)
-  const [balance, setBalance] = useState('0')
+  const balance = useTokenBalance(props.tokenAddress, account)
   const tokenSymbol = useTokenSymbol(props.tokenAddress ?? "")
   const [withdrawStatus, setWithdrawStatus] = useState(IDLE)
   const {info, refresh: fetchChannelInfo} = useActiveChannelInfo()
-
-  const refresh = useCallback(async () => {
-    if (!token) return
-    try {
-      const res = await token.balanceOf(account ?? ZERO_ADDRESS)
-      setBalance(formatNumber(parseToBigNumber(res).shiftedBy(-18)))
-    }catch (e){
-      setBalance('NaN')
-    }
-  }, [token, account])
-
-  setImmediate(refresh, 3000)
 
   const handleWithdraw = async () => {
     if (!nestOpenPlatform) return
@@ -275,7 +251,7 @@ const WithdrawPopover: FC<PopverProps> = ({...props}) => {
   )
 }
 
-const WithdrawFeePopover: FC<PopverProps> = ({...props}) => {
+const WithdrawFeePopover: FC<PopoverProps> = ({...props}) => {
   const {chainId, account} = useActiveWeb3React()
   const activeChannelId = useRecoilValue(activeChannelIdAtom)
   const nestOpenPlatform = useNestOpenPlatformContract(NEST_OPEN_PLATFORM_ADDRESS[chainId ?? 1], true)
